pyproject.toml
[project]
name = "flux2-mlx"
version = "0.1.0"
description = "FLUX.2 image generation for Apple Silicon using MLX"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "FLUX.2 MLX Contributors"}]
dependencies = [
    "mlx>=0.22.0",
    "tokenizers>=0.15.0",
    "jinja2>=3.0.0",
    "Pillow>=9.0.0",
    "numpy>=1.24.0",
    "huggingface_hub>=0.20.0",
]

[project.scripts]
flux2-mlx = "flux2_mlx.generate:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
